+++ src/main/java/io/github/astrapi69/test/base/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides a base class for tests
+ */
+package io.github.astrapi69.test.base;