+++ src/main/java/io/github/astrapi69/test/object/annotation/foobar/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides annotated classes that can be used unit test classes
+ */
+package io.github.astrapi69.test.object.annotation.foobar;