+++ src/main/java/io/github/astrapi69/test/object/classes/inner/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides utility classes for show the effects of inner in outer classes
+ */
+package io.github.astrapi69.test.object.classes.inner;