+++ src/main/java/io/github/astrapi69/test/object/date/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides decorator classes for date objects
+ */
+package io.github.astrapi69.test.object.date;