+++ src/main/java/io/github/astrapi69/test/object/factory/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package provides a factory class for generate all test objects that can be used unit test
+ * classes
+ */
+package io.github.astrapi69.test.object.factory;