+++ src/main/java/io/github/astrapi69/test/object/generic/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides test dao classes that can be used unit test classes
+ */
+package io.github.astrapi69.test.object.generic;