+++ src/main/java/io/github/astrapi69/test/object/type/hierarchy/shape/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides shape classes in a hierarchy
+ */
+package io.github.astrapi69.test.object.type.hierarchy.shape;