+++ src/main/java/module-info.java
@@ -1,17 +1,17 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
  * including without limitation the rights to use, copy, modify, merge, publish, distribute,
  * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in all copies or
  * substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
  * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
  * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
@@ -22,17 +22,15 @@ module test.object.main
 {
 	requires java.sql;
 
-	exports io.github.astrapi69;
-	exports io.github.astrapi69.classes.inner;
-	exports io.github.astrapi69.date;
-	exports io.github.astrapi69.test.instance;
-	exports io.github.astrapi69.test.message;
+	exports io.github.astrapi69.test.base;
 	exports io.github.astrapi69.test.object;
 	exports io.github.astrapi69.test.object.annotation;
 	exports io.github.astrapi69.test.object.annotation.classtype;
 	exports io.github.astrapi69.test.object.annotation.foobar;
 	exports io.github.astrapi69.test.object.annotation.interfacetype;
 	exports io.github.astrapi69.test.object.auth;
+	exports io.github.astrapi69.test.object.classes.inner;
+	exports io.github.astrapi69.test.object.date;
 	exports io.github.astrapi69.test.object.enumeration;
 	exports io.github.astrapi69.test.object.exception;
 	exports io.github.astrapi69.test.object.factory;