+++ src/test/java/io/github/astrapi69/test/object/date/SqlTimestampDecoratorTest.java
@@ -1,0 +1,69 @@
+/**
+ * The MIT License
+ * <p>
+ * Copyright (C) 2015 Asterios Raptis
+ * <p>
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ * <p>
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ * <p>
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.test.object.date;
+
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+
+import java.sql.Timestamp;
+import java.util.Date;
+
+import org.junit.jupiter.api.Disabled;
+import org.junit.jupiter.api.Test;
+import org.meanbean.test.BeanTester;
+
+import io.github.astrapi69.test.object.date.SqlTimestampDecorator;
+
+/**
+ * The unit test class for the class {@link SqlTimestampDecorator}.
+ */
+public class SqlTimestampDecoratorTest
+{
+
+	/**
+	 * Test method for {@link SqlTimestampDecorator} constructors
+	 */
+	@Test
+	public final void testConstructors()
+	{
+		SqlTimestampDecorator model = new SqlTimestampDecorator();
+		assertNotNull(model);
+		model = new SqlTimestampDecorator(new Timestamp(new Date().getTime()));
+		assertNotNull(model);
+		model = SqlTimestampDecorator.builder().build();
+		assertNotNull(model);
+	}
+
+	/**
+	 * Test method for {@link SqlTimestampDecorator}
+	 */
+	@Test
+	@Disabled
+	public void testWithBeanTester()
+	{
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(SqlTimestampDecorator.class);
+	}
+
+}